+++ tests/testthat/test-gelmansim.R
@@ -19,7 +19,7 @@ y1 <- y123.dat[,1]
 M1 <- glm (y1 ~ x1)
 
 cases <- data.frame(x1 = seq(-2, 2, by=0.1))
-sim.results <- gelmansim(mod = M1, newdata = cases, n.sims=200, na.omit=TRUE)
+sim.results <- suppress_deprecation(gelmansim(mod = M1, newdata = cases, n.sims=200, na.omit=TRUE))
 
 
 test_that("returned dataframe is correct size", {
@@ -43,7 +43,7 @@ M2 <- glm(y1 ~ x1 + group, data = dat)
 cases <- expand.grid(x1 = seq(-2, 2, by = 0.1), 
                      group = seq(1, 14, by = 2))
 
-sim.results <- gelmansim(M2, newdata=cases, n.sims=200, na.omit=TRUE)
+sim.results <- suppress_deprecation(gelmansim(M2, newdata=cases, n.sims=200, na.omit=TRUE))
 
 test_that("returned dataframe is correct size", {
   expect_equal(dim(sim.results)[1], nrow(cases))
@@ -67,7 +67,7 @@ cases <- expand.grid(x1 = seq(-2, 2, by=0.1),
                      group=seq(1, 14, by=2))
 cases$group <- factor(cases$group)
 
-sim.results <- gelmansim(M3, newdata=cases, n.sims=200, na.omit=TRUE)
+sim.results <- suppress_deprecation(gelmansim(M3, newdata=cases, n.sims=200, na.omit=TRUE))
 
 test_that("returned dataframe is correct size", {
   expect_equal(dim(sim.results)[1], nrow(cases))
@@ -91,7 +91,7 @@ cases <- expand.grid(x1 = seq(-2, 2, by=0.1),
                      group=seq(1, 14, by=2))
 cases$group <- factor(cases$group)
 
-sim.results <- gelmansim(M4, newdata=cases, n.sims=200, na.omit=TRUE)
+sim.results <- suppress_deprecation(gelmansim(M4, newdata=cases, n.sims=200, na.omit=TRUE))
 
 
 test_that("returned dataframe is correct size", {